feat(haskell): graph bfs and dfs searches
